requirements.txt
openai
python-dotenv
google-generativeai
numpy

#matplotlib
#plotly
#scikit-learn
#pandas